# Conflicts:
#   detecht_backend/detecht_api/models.py

+++ detecht_backend/detecht_api/detecht_converter/pdf_converter.py
@@ -1,32 +1,58 @@
 import PyPDF2
 import json
-from os import listdir
-from os.path import isfile, join
+import os
+import re
 
-# get the PDF path and read the file
-file = "BOAML_report_2017.pdf"
-read_pdf = PyPDF2.PdfFileReader(file, strict=False)
+def convert_pdf_to_json():
 
-# get the read object's meta info
-pdf_meta = read_pdf.getDocumentInfo()
-all_pages = {}
-all_pages["meta"] = {}
+    # get all PDFs  in the path
+    path = "detecht_api/static/pdf"
+    with os.scandir(path) as it:
+        for entry in it:
+            if entry.name.endswith(".pdf") and entry.is_file():
+                print("Currently scanning: " + entry.name)
+                file = entry.path
 
-for meta, value in pdf_meta.items():
-    all_pages["meta"][meta] = value
+                # read the current pdf
+                read_pdf = PyPDF2.PdfFileReader(file, strict=False)
+                all_pages = {}
 
-# iterate the page numbers
-for page in range(read_pdf.getNumPages()):
-    data = read_pdf.getPage(page)
-    page_mode = read_pdf.getPageMode()
-    page_text = data.extractText()
-    all_pages[page] = page_text
+                # iterate all pages
+                all_text = ""
+                for page in range(read_pdf.getNumPages()):
+                    data = read_pdf.getPage(page)
+                    page_text = data.extractText()
+                    modified_text = re.sub("\n", "", page_text)
+                    all_text = all_text + modified_text
 
+                all_pages["All_text"] = all_text
 
-# create a JSON string from the dictionary
-json_data = json.dumps(all_pages)
-# print ("\nJSON:", json_data)
+                # creates a JSON file
+                filename = entry.name
+                filename = filename[:-4]
+                filepath = 'detecht_api/static/json/' + filename + '.json'
+                with open(filepath,'w+') as outfile:
+                    json.dump(all_pages, outfile)
 
+def pdf_to_json(pdf_name):
 
-with open('test.json', 'w') as outfile:
-    json.dump(all_pages, outfile)
+    # get all PDFs  in the path
+    path = "detecht_api/static/pdf/"
+    with open(path + pdf_name, 'rb') as file:
+
+        # read the current pdf
+        read_pdf = PyPDF2.PdfFileReader(file, strict=False)
+        all_pages = {}
+
+        # iterate all pages
+        all_text = ""
+        for page in range(read_pdf.getNumPages()):
+            data = read_pdf.getPage(page)
+            page_text = data.extractText()
+            modified_text = re.sub("\n", "", page_text)
+            all_text = all_text + modified_text
+
+        all_pages["All_text"] = all_text
+
+        return all_text
+
